解决了编译时的警告

+++ kernel/swtimer.c
@@ -22,7 +22,7 @@
 
 typedef struct swtimer
 {
-    uint32_t timer_id;
+    uint64_t timer_id;
     uint64_t period; 
     int8_t status;//-1:销毁，0:暂停，1:开启
     uint64_t tick;
@@ -90,7 +90,7 @@ void swtimer_check()
         }
     }
 }
-
+ 
 /***************************************************************
  * @description: 创建一个软件定时器
  * @param void (*timer_task)() [in]: 定时器任务
@@ -104,10 +104,10 @@ swtimer_t* swtimer_create(void (*timer_task)(),uint64_t period,uint8_t mode)
     swtimer_t *new_timer = page_alloc(1);        
     if((swtimer_t*)NULL_PTR == new_timer) return NULL_PTR;
  
-    switch ((uint32_t)swtimer_head)
+    switch ((uint64_t)swtimer_head)
     {
-        case (uint32_t)NULL_PTR:
-            static uint32_t id = 0;
+        case (uint64_t)NULL_PTR:
+            static uint64_t id = 0;
             swtimer_head = page_alloc(1); 
             INIT_LIST_HEAD(&swtimer_head->swtimer_node);
         default:
